@@ -18,7 +18,10 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"io"
 	"net"
+	"os"
+	"sync"
 	"testing"
 	"time"
 
@@ -29,13 +32,21 @@ import (
 	"github.com/openconfig/gribigo/testcommon"
 	"go.uber.org/atomic"
 	"google.golang.org/grpc"
+	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/credentials"
+	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/encoding/prototext"
 	"google.golang.org/protobuf/testing/protocmp"
 
 	aftpb "github.com/openconfig/gribi/v1/proto/gribi_aft"
 	spb "github.com/openconfig/gribi/v1/proto/service"
 )
+
+func TestMain(m *testing.M) {
+	debug = true
+	defer func() { debug = false }()
+	os.Exit(m.Run())
+}
 
 func TestHandleParams(t *testing.T) {
 	tests := []struct {
@@ -162,12 +173,17 @@ func TestQ(t *testing.T) {
 			if err != nil {
 				t.Fatalf("cannot create client, %v", err)
 			}
+			doneCh := make(chan struct{})
 			if tt.inSending {
 				c.qs.sending = atomic.NewBool(tt.inSending)
 				// avoid test deadlock by emptying the queue if we're sending.
 				go func() {
 					for {
-						<-c.qs.modifyCh
+						select {
+						case <-c.qs.modifyCh:
+						case <-doneCh:
+							return
+						}
 					}
 				}()
 			}
@@ -178,6 +194,7 @@ func TestQ(t *testing.T) {
 			if diff := cmp.Diff(c.qs.sendq, tt.wantQ, protocmp.Transform()); diff != "" {
 				t.Fatalf("did not get expected send queue, %s", diff)
 			}
+			close(doneCh)
 		})
 	}
 }
@@ -1022,6 +1039,66 @@ func TestGet(t *testing.T) {
 			}},
 		},
 	}, {
+		desc: "single IPv6 entry in Server - specific NI",
+		inOperations: []*spb.AFTOperation{{
+			NetworkInstance: server.DefaultNetworkInstanceName,
+			Entry: &spb.AFTOperation_Ipv6{
+				Ipv6: &aftpb.Afts_Ipv6EntryKey{
+					Prefix:    "2001:db8::/32",
+					Ipv6Entry: &aftpb.Afts_Ipv6Entry{},
+				},
+			},
+		}},
+		inGetRequest: &spb.GetRequest{
+			NetworkInstance: &spb.GetRequest_Name{
+				Name: server.DefaultNetworkInstanceName,
+			},
+			Aft: spb.AFTType_ALL,
+		},
+		wantResponse: &spb.GetResponse{
+			Entry: []*spb.AFTEntry{{
+				NetworkInstance: server.DefaultNetworkInstanceName,
+				Entry: &spb.AFTEntry_Ipv6{
+					Ipv6: &aftpb.Afts_Ipv6EntryKey{
+						Prefix:    "2001:db8::/32",
+						Ipv6Entry: &aftpb.Afts_Ipv6Entry{},
+					},
+				},
+			}},
+		},
+	}, {
+		desc: "single MPLS entry in Server - specific NI",
+		inOperations: []*spb.AFTOperation{{
+			NetworkInstance: server.DefaultNetworkInstanceName,
+			Entry: &spb.AFTOperation_Mpls{
+				Mpls: &aftpb.Afts_LabelEntryKey{
+					Label: &aftpb.Afts_LabelEntryKey_LabelUint64{
+						LabelUint64: 42,
+					},
+					LabelEntry: &aftpb.Afts_LabelEntry{},
+				},
+			},
+		}},
+		inGetRequest: &spb.GetRequest{
+			NetworkInstance: &spb.GetRequest_Name{
+				Name: server.DefaultNetworkInstanceName,
+			},
+			Aft: spb.AFTType_ALL,
+		},
+		wantResponse: &spb.GetResponse{
+			Entry: []*spb.AFTEntry{{
+				NetworkInstance: server.DefaultNetworkInstanceName,
+				Entry: &spb.AFTEntry_Mpls{
+					Mpls: &aftpb.Afts_LabelEntryKey{
+						Label: &aftpb.Afts_LabelEntryKey_LabelUint64{
+							LabelUint64: 42,
+						},
+						LabelEntry: &aftpb.Afts_LabelEntry{},
+					},
+				},
+			}},
+		},
+	}, {
 		desc: "multiple entries - single NI",
 		inOperations: []*spb.AFTOperation{{
 			NetworkInstance: server.DefaultNetworkInstanceName,
@@ -1513,6 +1590,7 @@ func TestFlush(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.desc, func(t *testing.T) {
+			defer tt.inClient.Close()
 			creds, err := credentials.NewServerTLSFromFile(testcommon.TLSCreds())
 			if err != nil {
 				t.Fatalf("cannot load TLS credentials, got err: %v", err)
@@ -1608,7 +1686,13 @@ func TestServerModifyIntegration(t *testing.T) {
 				return fmt.Errorf("Connect(): cannot connect to server, %v", err)
 			}
 
-			c.Q(&spb.ModifyRequest{})
+			c.Q(&spb.ModifyRequest{
+				Params: &spb.SessionParameters{
+					AckType:     spb.SessionParameters_RIB_ACK,
+					Redundancy:  spb.SessionParameters_SINGLE_PRIMARY,
+					Persistence: spb.SessionParameters_PRESERVE,
+				},
+			})
 			c.StartSending()
 
 			if err := c.AwaitConverged(ctx); err != nil {
@@ -1719,4 +1803,521 @@ func TestServerModifyIntegration(t *testing.T) {
 		})
 	}
 
+}
+
+type disconnectMode int64
+
+const (
+	_ disconnectMode = iota
+	EOF
+	PauseEOF
+	PauseError
+	LongPauseError
+	NoError
+	PauseNoError
+)
+
+func (d disconnectMode) String() string {
+	switch d {
+	case EOF:
+		return "EOF"
+	case PauseEOF:
+		return "EOF after pause"
+	case PauseError:
+		return "Server error after pause"
+	case LongPauseError:
+		return "Server error after 100s pause"
+	case NoError:
+		return "nil error"
+	case PauseNoError:
+		return "nil error after pause"
+	default:
+		return fmt.Sprintf("%d", d)
+	}
+}
+
+type disconnectingGRIBI struct {
+	*spb.UnimplementedGRIBIServer
+	mode disconnectMode
+	// Set to true if we should also read the channel.
+	receiver bool
+
+	mu   sync.Mutex
+	msgs []*spb.ModifyRequest
+}
+
+func (d *disconnectingGRIBI) Modify(stream spb.GRIBI_ModifyServer) error {
+	var wg sync.WaitGroup
+
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		for {
+			m, err := stream.Recv()
+			if err != nil {
+				return
+			}
+			d.mu.Lock()
+			d.msgs = append(d.msgs, m)
+			d.mu.Unlock()
+		}
+	}()
+
+	var err error
+	switch d.mode {
+	case EOF:
+		err = io.EOF
+	case PauseEOF:
+		time.Sleep(2 * time.Second)
+		err = io.EOF
+	case PauseError:
+		time.Sleep(2 * time.Second)
+		err = status.Errorf(codes.ResourceExhausted, "yawn!")
+	case LongPauseError:
+		time.Sleep(100 * time.Second)
+		err = status.Errorf(codes.ResourceExhausted, "slept 100 seconds")
+	case NoError:
+	case PauseNoError:
+		time.Sleep(2 * time.Second)
+	}
+	if d.receiver {
+		wg.Wait()
+	}
+	return err
+}
+
+func (d *disconnectingGRIBI) Flush(_ context.Context, _ *spb.FlushRequest) (*spb.FlushResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "unimplemented")
+}
+
+func (d *disconnectingGRIBI) Get(_ *spb.GetRequest, _ spb.GRIBI_GetServer) error {
+	return status.Errorf(codes.Unimplemented, "unimplemented")
+}
+
+func newDisconnectingFake(t testing.TB, mode disconnectMode) (*disconnectingGRIBI, string, func()) {
+	creds, err := credentials.NewServerTLSFromFile(testcommon.TLSCreds())
+	if err != nil {
+		t.Fatalf("cannot load TLS credentials, got err: %v", err)
+	}
+	srv := grpc.NewServer(grpc.Creds(creds))
+	if err != nil {
+		t.Fatalf("cannot create server, err: %v", err)
+	}
+	s := &disconnectingGRIBI{mode: mode}
+	spb.RegisterGRIBIServer(srv, s)
+
+	l, err := net.Listen("tcp", "localhost:0")
+	if err != nil {
+		t.Fatalf("cannot listen, err: %v", err)
+	}
+
+	go srv.Serve(l)
+	return s, l.Addr().String(), srv.Stop
+}
+
+func TestDone(t *testing.T) {
+	tests := []struct {
+		desc                 string
+		inMode               disconnectMode
+		inReconnects         int
+		inTimeoutSeconds     int
+		inWaitTimeoutSeconds int
+		wantDone             bool
+	}{{
+		desc:         "EOF returned",
+		inMode:       EOF,
+		inReconnects: 5,
+		wantDone:     true,
+	}, {
+		desc:         "pause",
+		inMode:       PauseError,
+		inReconnects: 1,
+		wantDone:     true,
+	}, {
+		desc:                 "inner context is cancelled",
+		inMode:               LongPauseError,
+		inTimeoutSeconds:     10,
+		inWaitTimeoutSeconds: 1,
+		inReconnects:         1,
+		wantDone:             true,
+	}, {
+		desc:         "no error returned",
+		inMode:       NoError,
+		inReconnects: 1,
+		wantDone:     true,
+	}}
+
+	for _, tt := range tests {
+		t.Run(tt.desc, func(t *testing.T) {
+
+			_, addr, stopSrv := newDisconnectingFake(t, tt.inMode)
+			defer stopSrv()
+
+			c, err := New()
+			if err != nil {
+				t.Fatalf("cannot create client, err: %v", err)
+			}
+
+			dur := 5 * time.Second
+			if tt.inTimeoutSeconds != 0 {
+				dur = time.Duration(tt.inTimeoutSeconds) * time.Second
+			}
+
+			dctx, cancel := context.WithTimeout(context.Background(), dur)
+			defer cancel()
+			if err := c.Dial(dctx, addr); err != nil {
+				t.Fatalf("c.Dial(_, %s): cannot dial fake server, got err: %v", addr, err)
+			}
+
+			for i := 0; i <= tt.inReconnects; i++ {
+				t.Logf("running reconnect %d for %s\n", i, tt.desc)
+				var (
+					wg     sync.WaitGroup
+					retErr error
+				)
+
+				innerDur := 5 * time.Second
+				if tt.inWaitTimeoutSeconds != 0 {
+					innerDur = time.Duration(tt.inWaitTimeoutSeconds) * time.Second
+				}
+
+				waitCtx, cancel := context.WithTimeout(dctx, innerDur)
+				defer cancel()
+
+				if err := c.Connect(waitCtx); err != nil {
+					retErr = fmt.Errorf("Connect(): cannot connect to server, %v", err)
+				}
+
+				c.Q(&spb.ModifyRequest{})
+				c.StartSending()
+
+				var got bool
+				wg.Add(1)
+				go func(ctx context.Context) {
+					defer wg.Done()
+					select {
+					case <-c.Done():
+						got = true
+						return
+					case <-ctx.Done():
+						// This case only happens if the overall test times out.
+					}
+				}(dctx)
+
+				wg.Wait()
+				if retErr != nil {
+					t.Fatalf("did not connect to server, got: %v, want: nil", retErr)
+				}
+				if got != tt.wantDone {
+					t.Fatalf("did not get done signal, got: %v, want: %v", got, tt.wantDone)
+				}
+				c.StopSending()
+				c.Reset()
+			}
+		})
+	}
+}
+
+func TestReconnect(t *testing.T) {
+	// How many runs to do for unstable tests. Note that some tests have relatively long runtime
+	// so this might result in test target timeouts if too high.
+	manyRuns := 30
+	tests := []struct {
+		desc           string
+		inModes        []disconnectMode // Must use Pause modes if results are expected.
+		inReconnects   int
+		inClientOpts   []Opt
+		inMsgs         []*spb.ModifyRequest
+		inMsgFn        func(int) []*spb.ModifyRequest
+		inCallStop     bool
+		inQBeforeStart bool
+		inRunTime      time.Duration
+		inSkipReset    bool
+		inStopServer   bool
+		wantMsgs       []*spb.ModifyRequest
+		wantConnectErr bool
+	}{{
+		desc:         "reconnect - stop sending not called",
+		inModes:      []disconnectMode{PauseNoError, PauseError, PauseEOF},
+		inReconnects: 1,
+		inClientOpts: []Opt{
+			PersistEntries(),
+		},
+		inMsgs: []*spb.ModifyRequest{{
+			ElectionId: &spb.Uint128{Low: 1},
+		}},
+		wantMsgs: []*spb.ModifyRequest{
+			// Must get parameters before any other message.
+			{Params: &spb.SessionParameters{Persistence: spb.SessionParameters_PRESERVE}},
+			{ElectionId: &spb.Uint128{Low: 1}},
+			{Params: &spb.SessionParameters{Persistence: spb.SessionParameters_PRESERVE}},
+			{ElectionId: &spb.Uint128{Low: 1}},
+		},
+	}, {
+		desc:         "reconnect - stop sending called",
+		inModes:      []disconnectMode{PauseNoError, PauseError, EOF},
+		inReconnects: 1,
+		inClientOpts: []Opt{
+			PersistEntries(),
+		},
+		inMsgs: []*spb.ModifyRequest{{
+			ElectionId: &spb.Uint128{Low: 1},
+		}},
+		inCallStop: true,
+		wantMsgs: []*spb.ModifyRequest{
+			// Must get parameters before any other message.
+			{Params: &spb.SessionParameters{Persistence: spb.SessionParameters_PRESERVE}},
+			{ElectionId: &spb.Uint128{Low: 1}},
+			{Params: &spb.SessionParameters{Persistence: spb.SessionParameters_PRESERVE}},
+			{ElectionId: &spb.Uint128{Low: 1}},
+		},
+	}, {
+		desc:         "reconnect - client qs before calling start",
+		inModes:      []disconnectMode{PauseNoError, PauseError, PauseEOF},
+		inReconnects: 1,
+		inClientOpts: []Opt{
+			PersistEntries(),
+		},
+		inMsgs: []*spb.ModifyRequest{{
+			ElectionId: &spb.Uint128{Low: 1},
+		}},
+		inQBeforeStart: true,
+		wantMsgs: []*spb.ModifyRequest{
+			// Must get parameters before any other message.
+			{Params: &spb.SessionParameters{Persistence: spb.SessionParameters_PRESERVE}},
+			{ElectionId: &spb.Uint128{Low: 1}},
+			{Params: &spb.SessionParameters{Persistence: spb.SessionParameters_PRESERVE}},
+			{ElectionId: &spb.Uint128{Low: 1}},
+		},
+	}, {
+		desc:         "reconnect - client qs before calling start but has stopped",
+		inModes:      []disconnectMode{PauseNoError, PauseError, PauseEOF},
+		inReconnects: 1,
+		inClientOpts: []Opt{
+			PersistEntries(),
+		},
+		inMsgs: []*spb.ModifyRequest{{
+			ElectionId: &spb.Uint128{Low: 1},
+		}},
+		inCallStop:     true,
+		inQBeforeStart: true,
+		wantMsgs: []*spb.ModifyRequest{
+			// Must get parameters before any other message.
+			{Params: &spb.SessionParameters{Persistence: spb.SessionParameters_PRESERVE}},
+			{ElectionId: &spb.Uint128{Low: 1}},
+			{Params: &spb.SessionParameters{Persistence: spb.SessionParameters_PRESERVE}},
+			{ElectionId: &spb.Uint128{Low: 1}},
+		},
+	}, {
+		desc:         "very unstable server",
+		inModes:      []disconnectMode{PauseNoError, PauseEOF, PauseError},
+		inReconnects: manyRuns,
+		inClientOpts: []Opt{
+			PersistEntries(),
+			ElectedPrimaryClient(&spb.Uint128{Low: 42}),
+		},
+		inMsgFn: func(attempt int) []*spb.ModifyRequest {
+			return []*spb.ModifyRequest{
+				{ElectionId: &spb.Uint128{Low: uint64(42 + attempt)}},
+			}
+		},
+		inRunTime: time.Duration(manyRuns) * 2 * time.Second,
+		wantMsgs: func() []*spb.ModifyRequest {
+			msgs := []*spb.ModifyRequest{}
+			for i := 0; i <= manyRuns; i++ {
+				msgs = append(msgs,
+					[]*spb.ModifyRequest{
+						{Params: &spb.SessionParameters{Persistence: spb.SessionParameters_PRESERVE, Redundancy: spb.SessionParameters_SINGLE_PRIMARY}},
+						{ElectionId: &spb.Uint128{Low: 42}},
+						{ElectionId: &spb.Uint128{Low: uint64(42 + i)}},
+					}...)
+			}
+			return msgs
+		}(),
+	}, {
+		desc:         "very unstable server - don't check results",
+		inModes:      []disconnectMode{NoError, EOF},
+		inReconnects: manyRuns,
+		inClientOpts: []Opt{
+			PersistEntries(),
+			ElectedPrimaryClient(&spb.Uint128{Low: 42}),
+		},
+		inMsgFn: func(attempt int) []*spb.ModifyRequest {
+			return []*spb.ModifyRequest{
+				{ElectionId: &spb.Uint128{Low: uint64(42 + attempt)}},
+			}
+		},
+		inRunTime: time.Duration(manyRuns) * 2 * time.Second,
+	}, {
+		desc:         "server that isn't even listening",
+		inModes:      []disconnectMode{EOF}, // Doesn't matter, we are not going to dial it.
+		inReconnects: 10,                    // More than the buffer depth of modifyCh.
+		inRunTime:    5 * time.Second,
+		inStopServer: true,
+	}}
+
+	for _, tt := range tests {
+		for _, mode := range tt.inModes {
+			t.Run(fmt.Sprintf("%s - mode %s", tt.desc, mode), func(t *testing.T) {
+				server, addr, stop := newDisconnectingFake(t, mode)
+				server.receiver = true
+				defer stop()
+
+				c, err := New(tt.inClientOpts...)
+				if err != nil {
+					t.Fatalf("cannot create client, err: %v", err)
+				}
+				to := tt.inRunTime
+				if to.Seconds() < 2.0 {
+					to = 2 * time.Second
+				}
+				dctx, cancel := context.WithTimeout(context.Background(), to)
+				defer cancel()
+
+				if err := c.Dial(dctx, addr); err != nil {
+					t.Fatalf("c.Dial(_, %s): cannot dial fake server, got err: %v", addr, err)
+				}
+
+				for i := 0; i <= tt.inReconnects; i++ {
+					if tt.inStopServer && i > 0 {
+						// Kill the server so that its no longer listening.
+						stop()
+					}
+					t.Logf("reconnecting to server (mode: %s), attempt: %d", mode, i)
+					var (
+						wg     sync.WaitGroup
+						retErr error
+					)
+
+					select {
+					case <-dctx.Done():
+						t.Fatalf("dial context timed out, %v", err)
+					default:
+					}
+
+					waitCtx, stopWait := context.WithTimeout(dctx, 200*time.Millisecond)
+					defer stopWait()
+
+					if err := c.Connect(waitCtx); err != nil {
+						switch {
+						case tt.inStopServer && i == 0, !tt.inStopServer:
+							retErr = fmt.Errorf("Connect(): cannot connect to server, %v", err)
+						default:
+						}
+					}
+
+					runMsg := tt.inMsgs
+					if tt.inMsgFn != nil {
+						runMsg = tt.inMsgFn(i)
+					}
+
+					if tt.inQBeforeStart {
+						for _, m := range runMsg {
+							c.Q(m)
+						}
+					}
+
+					c.StartSending()
+					if tt.inCallStop {
+						defer c.StopSending()
+					}
+
+					if !tt.inQBeforeStart {
+						for _, m := range runMsg {
+							c.Q(m)
+						}
+					}
+
+					var reconnected bool
+					wg.Add(1)
+					go func(ctx context.Context) {
+						defer wg.Done()
+						select {
+						case <-c.Done():
+							reconnected = true
+							return
+						case <-ctx.Done():
+							reconnected = true
+							return
+						}
+					}(waitCtx)
+
+					wg.Wait()
+					// Sanity check - we should have no error, and have the server go away.
+					if retErr != nil {
+						t.Fatalf("did not connect to server, got: %v, want: nil", retErr)
+					}
+					if !reconnected {
+						t.Fatalf("did not reconnect to server, got: %v, want: true", reconnected)
+					}
+					c.Reset()
+				}
+
+				if tt.wantMsgs != nil {
+					// Now check the messages that we saw at the server.
+					server.mu.Lock()
+					defer server.mu.Unlock()
+					if diff := cmp.Diff(server.msgs, tt.wantMsgs, protocmp.Transform()); diff != "" {
+						t.Fatalf("did not get expected messages, diff(-got,+want):\n%s", diff)
+					}
+				}
+			})
+		}
+	}
+}
+
+func TestReset(t *testing.T) {
+	newC := func() *Client {
+		c, _ := New()
+		return c
+	}
+	tests := []struct {
+		desc     string
+		inClient *Client
+	}{{
+		desc:     "no fields set",
+		inClient: newC(),
+	}, {
+		desc: "set fields",
+		inClient: func() *Client {
+			c, _ := New()
+			c.sendErr = []error{errors.New("fish")}
+			c.readErr = []error{errors.New("chips")}
+			c.qs.sendq = []*spb.ModifyRequest{{}}
+			c.qs.pendq = &pendingQueue{
+				Ops: map[uint64]*PendingOp{1: {}},
+			}
+			c.qs.resultq = []*OpResult{{}}
+			return c
+		}(),
+	}}
+
+	for _, tt := range tests {
+		t.Run(tt.desc, func(t *testing.T) {
+			// Setup, make sure it looks like our goroutines are started.
+			tt.inClient.sendExitCh = make(chan struct{}, 1)
+			tt.inClient.sendExitCh <- struct{}{}
+			close(tt.inClient.sendExitCh)
+
+			tt.inClient.Reset()
+			c := tt.inClient
+			// Don't compare the whole struct because there are unexported fields and it is not for public consumption.
+
+			if len(c.sendErr) != 0 {
+				t.Errorf("unexpected send errors, got: %d, want: 0", len(c.sendErr))
+			}
+			if len(c.readErr) != 0 {
+				t.Errorf("unexpected read errors, got: %d, want: 0", len(c.readErr))
+			}
+			if len(c.qs.sendq) != 0 {
+				t.Errorf("unexpected send queue got: %d, want: 0", len(c.qs.sendq))
+			}
+			if len(c.qs.pendq.Ops) != 0 {
+				t.Errorf("unexpected pending queue, got: %d, want: 0", len(c.qs.pendq.Ops))
+			}
+			if len(c.qs.resultq) != 0 {
+				t.Errorf("unexpected results queue, got: %d, want: 0", len(c.qs.resultq))
+			}
+		})
+	}
 }
